Somewhat working ISRs

+++ kernel/types.h
@@ -8,7 +8,7 @@ typedef unsigned short uint16_t;
 typedef          short int16_t;
 typedef unsigned char  uint8_t;
 typedef          char  int8_t;
-typedef unsigned long int	uintptr_t;
+typedef unsigned long int	uint64_t;
 
 #define NULL 0
